t020 test 9: Fix spurious failure due to user's git config
It seems that `git reflog` uses the user's `log.decorate` configuration
setting, so a user who sets that to, e.g., `short` will see a spurious
failure due to the branch names being printed.

Explicitly specify `--no-decorate` to produce consistent, minimal output.

This isn't documented anywhere in the manpage for git 1.9.1, so it's
not clear where and when this really applies. We really need a test
framework that can test this under multiple versions of git.