t/*: Fix shebang lines
Two were not valid because they started with just `#` insted of `#!`.
Also they assumed Bash was in `/bin/`. Change these all to the
standard `#!/usr/bin/env bash`.

The shebangs are not actually used: the files aren't even marked
executable and they're not run directly but run by `prove`, where we
specify the interpreter to use. But these serve as documentation,
particularly for automated tools.

It happened that GitHub was identifying `t/40-commit.t` as Perl,
instead of Bash, and that file being a plurality of the code in the
repo Github indicated the project language as Perl in its summary
information. This fixes that.