t/10-test-lib: expect_fails assertion, test suite for test-lib
This adds a new assertion, `expect_fails` to expect that a
`test_equal` fails instead of passes. That's not done in a
totally elegant way, and the test framework, such as it is,
probably needs a rework if much more of this comes along.

We also add tests for test_equal.

This is all in the interests of expanding our assertions
slightly to deal with the broken tests under Git 2.17.1.