t/40-commit: Extract assert_branch function and abbreviate commits
All calls to git-show-ref now set --abbrev=12 to make things a bit
more short and sweet. If this fails under earlier versions of Git
that we need to support with tests, we'll have to back this out,
but it's been tested on 2.11.0 and 1.9.1.

The new assert_branch function in most cases takes the just the
git-show-ref output, but you can give it a branch name if you need to
get information for a branch other than the default $branch we put our
test commits on.