If local branch has a tracking branch, fast-forward to tracking branch head
This is almost invariably the desired behaviour and makes this a lot
more convenient to use when others are also committing to the release
branch.

We fail if we can't fast-forward; we could consider making this a
default option rather than the only behaviour.

This has pushed our hacked-together testing framework about as far as
it can conveniently go (or probably beyond). Any further significant
work on this probably wants a test framework rewrite, say, in Python.

And boy oh boy did we learn about `git update-ref`. (There's a comment
in the code.)

Reviewed-by: Nishant Rodrigues <nishantjr@gmail.com>