WIP t/40-commit: Test currently showing we don't fast-forward local to tracking
XXX We still do not have a test for aborting if we can't fast-foward
the local to the tip of tracking. See below for details.

Now all the remains is to implement the fast-forward of the local
branch up to the tip of the tracking branch.

This may be tricker than it seems:
* `git rebase` not only has no --ff-only option, but it uses and
  changes a local working copy, which we'd like to avoid, and we ought
  not need any rebasing anyway if we can directly fast-forward. (If we
  can't, we probably want to abort.)
* `git merge` works only with the current HEAD, which is not the branch
  we want to fast-forward.

So it seems like we want to figure out ourselves if the change would
be  a fast-forward and, if it would be, just use `git update-ref -m
... $ref $ref{upstream}` to point the local ref at its upstream. Maybe
we can use something that asks, "is $ref a direct ancestor of
$ref{upstream}"?

It looks as if Nishant had already done some work on this more than
two years ago; see njr/160114/fast-forward.