GIT_COMMIT_MESSAGE envvar hack
You can now set GIT_COMMIT_MESSAGE in the environment to override the
default commit message.

This is probably not a good hack, and not just because it has no
tests. Using environment variables runs the risk that they are
unwittingly set when the caller runs the program and she gets a commit
message she didn't want. This should be changed to be a command-line
parameter so that can't happen.

This may also be an issue with GIT_AUTHOR_NAME etc.; we need to look
in to this.

So for the moment we leave this on a branch, and undocumented, but
it's here for anybody who needs it and is happy with the risks.